refactor: modularize icons, models, and idb services
